ASoC: wcd9xxx-mbhc: Add PLUG_TYPE_HIGH_HPH to fast-detection path
Our hardware commonly mistakes certain multi-button headsets for
high-hph (ex. Apple EarPods), delaying detection by up to 2 seconds
and resulting in poor UX.

Set the fast-detection plug-type to headset when high-hph is detected to
reduce detection latency by up to 400% for most of these usecases. If this
assumption is wrong and the inserted plug is not for a headset, then we will
just defer to the correct-plug worker to fix the incorrect detection.

Signed-off-by: Sultanxda <sultanxda@gmail.com>