ASoC: wcd9xxx-mbhc: Fix discrepancy for PLUG_TYPE_GND_MIC_SWAP
When fast detection is used, the pt_gnd_mic_swap_cnt variable is never
incremented, creating a potential discrepancy in the correct-plug worker's
reliability. For fast detection, the goal with wcd9xxx_do_plug_correction()
is to do everything exactly the same as when fast detection isn't used,
with the exception of constantly reporting different plug types.

Signed-off-by: Sultanxda <sultanxda@gmail.com>