ASoC: wcd9xxx-mbhc: Prevent spurious headset button interrupts
Sometimes when plugging in a headset, a spurious button press will be
registered, causing buggy behavior such as the volume-down/volume-up
buttons being pressed an infinite amount of times. This causes the volume
level to either skyrocket to the maximum or the minimum, and stay stuck
there until the headset is removed.

Ignoring all headset button interrupts during the 750ms after plug
detection is finished prevents the spurious interrupts from taking over.

Signed-off-by: Sultanxda <sultanxda@gmail.com>