power: qpnp-charger: Ensure I2C bus is active when plugging in USB cable
When the device is power-collapsed and a USB cable is inserted, multiple
I2C errors are observed:
[ 3892.212584] bq24196 read i2c failed,reg_addr=0x1,rc=-1
[ 3892.212587] bq24196 read i2c failed,reg_addr=0x1,rc=-1
[ 3892.212601] bq24196 read i2c failed,reg_addr=0x5,rc=-1

These errors don't impact functionality, but it's better to fix them.

Delays have been added to the relevant sections of code where the errors are
observed (bq24196_wait_for_resume() is used as a dynamic delay), in order to
wait until the I2C bus is active. These delays are typically ~10ms.

Signed-off-by: Sultanxda <sultanxda@gmail.com>